+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `ng-cloak` to root
+- `ng-strict-di` to root
+- New config `compileProvider`
+
+### Changed
+- Styles
 
 ## [0.1.39]
 ### Added